+++ CHANGELOG.rst
@@ -18,7 +18,8 @@ Backward-incompatible changes:
 Deprecations:
 ^^^^^^^^^^^^^
 
-*none*
+- Deprecated ``OpenSSL.rand`` - callers should use ``os.urandom()`` instead.
+  `#658 <https://github.com/pyca/pyopenssl/pull/658>`_
 
 
 Changes: